+++ src/FunFair.CodeAnalysis/Helpers/Rules.cs
@@ -41,5 +41,6 @@ namespace FunFair.CodeAnalysis.Helpers
         public const string RuleDontUseGuidParse = @"FFS0037";
         public const string RuleRecordsShouldSpecifyDebuggerDisplay = @"FFS0038";
         public const string RuleOnlyOneTypeDefinedPerFile = @"FFS0039";
+        public const string RuleTypeShouldBeInAFileWithSameName = @"FFS0040";
     }
 }