+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated TeamCity.VSTest.TestAdapter to 1.0.27
+- FF-1429 - Updated Microsoft.VisualStudio.Threading.Analyzers to 17.0.63
 ### Removed
 ### Deployment Changes