+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - FF-1429 - Updated TeamCity.VSTest.TestAdapter to 1.0.27
 - FF-1429 - Updated Microsoft.VisualStudio.Threading.Analyzers to 17.0.63
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.30.0.37606
+- FF-1429 - Updated Microsoft.CodeAnalysis.Analyzers to 3.3.3
 ### Removed
 ### Deployment Changes