+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.32.0.39516
+- FF-1429 - Updated Microsoft.CodeAnalysis.CSharp.Workspaces to 4.0.1
 ### Removed
 ### Deployment Changes